refactor: move common componetn
